Revert "Move skill validation from PostToolUse to Stop hook"
This reverts commit d9941de9c1f1db5b6b5a66e4176870e23dc80b20.
