+++ plugins/power-pages/hooks/hooks.json
@@ -11,17 +11,6 @@
           }
         ]
       }
-    ],
-    "Stop": [
-      {
-        "hooks": [
-          {
-            "type": "command",
-            "command": "node \"${CLAUDE_PLUGIN_ROOT}/hooks/run-skill-stop-validation.js\"",
-            "timeout": 30
-          }
-        ]
-      }
     ]
   }
 }
